Issue #70 refactor: primary key 타입 변환
Int -> Long

+++ app/src/main/java/com/drunkenboys/calendarun/data/checkpoint/entity/CheckPoint.kt
@@ -18,8 +18,8 @@ import java.util.*
     ]
 )
 data class CheckPoint(
-    @PrimaryKey(autoGenerate = true) val id: Int = 0,
-    val calendarId: Int,
+    @PrimaryKey(autoGenerate = true) val id: Long = 0,
+    val calendarId: Long,
     val name: String,
     val date: Date
 )
